Creating a Singly Linked List and Usage of Nodes
This is an activity from my Data Structures and Algorithms course. Basically, this is my first time creating a Node class, wherein it is a requirement for creating a Singly Linked List. I let ChatGPT help me and looking up some syntaxes online. This was my first time ever creating a Singly Linked List and we never knew the syntax or the code needed to create this so I asked an AI to teach me and I also did some manual research. Here, I was able to learn how to create a Node class, how to link different nodes in order, and also how to display them in order. It is a simple example, it simulates some sort of music playlist or queue. Uploaded this for reference.